runtime: give the aarch64 outline-atomics helpers a 64-bit operand
The embedded compiler-rt typedefs `__bcrt_u64` as `unsigned long`, which
is 4 bytes on Windows LLP64. Every `*8_*` helper the generator macros
build over it -- ldadd8, ldclr8, ldeor8, ldset8, swp8, cas8 -- therefore
carried a 4-byte operand on windows-aarch64 while its name promises an
8-byte one, so a caller's upper half was neither read nor written.
Measured on windows-arm64: `__aarch64_ldadd8_acq_rel(1, &a)` over
a = 0x1000000000 returned 0 and left a = 0x1000000001, and
`__aarch64_swp8_acq_rel(0x3000000000, &b)` over b = 0x2000000000 returned
0 and left b unchanged. `unsigned long long` is 8 bytes on every target,
so LP64 codegen is unchanged.

`OUTLINE_ATOMICS_SRC` hid this twice over: it typedefs its own `u64` as
`unsigned long`, so both sides shrank together, and its 8-byte case used
values that fit in 32 bits. It now uses `long long` and operands that
carry a value in the upper half, plus a swp8 case, and fails against the
previous helper widths.

+++ libc/lib/compiler_rt.c
@@ -27,7 +27,9 @@
 typedef unsigned char __bcrt_u8;
 typedef unsigned short __bcrt_u16;
 typedef unsigned int __bcrt_u32;
-typedef unsigned long __bcrt_u64;
+// `long` is 4 bytes on Windows LLP64, which would give every `*8_*`
+// helper a 4-byte operand; `long long` is 8 on every target.
+typedef unsigned long long __bcrt_u64;
 
 #define BCRT_LDADD(sz, T, ord, mo) \
     T __aarch64_ldadd##sz##_##ord(T v, T *p) { return __atomic_fetch_add(p, v, mo); }
